Update frontend.json
avg requests per server is a bit "weird" because we typically limit this and the impression of system usage is IMHO better visible by showing the sum of all active sessions for the backend at that time. this also allows showing different scaling effects when the number of backend servers changes.

BTW: I'm not sure why "metric" says "backend_current_queue" here.